More improvements to the testing script.
Tried to improve speed by only writing to files at the end of all the tests. This did not work so well; the bottleneck must be the amount of time all the test cases take.

Added an option to specify a maximum amount of cards to test.

Added an option to include (or not include) cards that cannot be solved.